fixed loading and broke out code
